updated anndata import for tabula muris
Changed this to import `read_h5ad` directly from `anndata` in case people do not have `Scanpy` installed.

I will play around with this in a minute. It would be cool to avoid the transposition when loading an `h5ad` file. As explained  [here](https://anndata.readthedocs.io/en/latest/anndata.AnnData.html), `AnnData` uses the `n_observations x n_variables` = `n_cells x n_genes` convention:
```
`AnnData` stores observations (samples) of variables (features) in the rows of a matrix. This is the convention of the modern classics of statistics [Hastie09] and machine learning [Murphy12], the convention of dataframes both in Python and R and the established statistics and machine learning packages in Python (statsmodels, scikit-learn).
```
But, sure, it's also no big problem to transpose after loading.